requirements.txt
ultralytics==8.0.196
streamlit==1.28.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
plotly==5.17.0
matplotlib==3.7.2
scipy==1.11.3
requests==2.31.0